API: change assignCertificationOfficer of repo
