+++ api/tests/integration/infrastructure/repositories/session-repository_test.js
@@ -452,54 +452,4 @@ describe('Integration | Repository | Session', function() {
     });
   });
 
-  describe('#assignCertificationOfficer', () => {
-    let sessionId;
-    let assignedCertificationOfficerId;
-
-    beforeEach(() => {
-      sessionId = databaseBuilder.factory.buildSession({ assignedCertificationOfficerId: null }).id;
-      assignedCertificationOfficerId = databaseBuilder.factory.buildUser().id;
-
-      return databaseBuilder.commit();
-    });
-
-    it('should return an updated Session domain object', async () => {
-      // when
-      const updatedSession = await sessionRepository.assignCertificationOfficer({ id: sessionId, assignedCertificationOfficerId });
-
-      // then
-      expect(updatedSession).to.be.an.instanceof(Session);
-      expect(updatedSession.id).to.deep.equal(sessionId);
-      expect(updatedSession.assignedCertificationOfficerId).to.deep.equal(assignedCertificationOfficerId);
-      expect(updatedSession.status).to.deep.equal(statuses.IN_PROCESS);
-    });
-
-    context('when assignedCertificationOfficerId provided does not exist', () => {
-
-      it('should return a Not found error', async () => {
-        // given
-        const unknownUserId = assignedCertificationOfficerId + 1;
-
-        // when
-        const error = await catchErr(sessionRepository.assignCertificationOfficer)({ id: sessionId, assignedCertificationOfficerId: unknownUserId });
-
-        // then
-        expect(error).to.be.instanceOf(NotFoundError);
-      });
-    });
-
-    context('when sessionId does not exist', () => {
-
-      it('should return a Not found error', async () => {
-        // given
-        const unknownSessionId = sessionId + 1;
-
-        // when
-        const error = await catchErr(sessionRepository.assignCertificationOfficer)({ id: unknownSessionId, assignedCertificationOfficerId });
-
-        // then
-        expect(error).to.be.instanceOf(NotFoundError);
-      });
-    });
-  });
 });
