+++ api/tests/unit/domain/usecases/assign-certification-officer-to-session_test.js
@@ -5,10 +5,10 @@ const { ObjectValidationError } = require('../../../../lib/domain/errors');
 describe('Unit | UseCase | assign-certification-officer-to-session', () => {
   let sessionId;
   let certificationOfficerId;
-  let sessionRepository;
+  let jurySessionRepository;
 
   beforeEach(() => {
-    sessionRepository = { assignCertificationOfficer: sinon.stub() };
+    jurySessionRepository = { assignCertificationOfficer: sinon.stub() };
   });
 
   context('when session id is not a number', () => {
@@ -18,7 +18,7 @@ describe('Unit | UseCase | assign-certification-officer-to-session', () => {
       sessionId = 'notANumber';
 
       // when
-      const error = await catchErr(assignCertificationOfficerToSession)({ sessionId, certificationOfficerId, sessionRepository });
+      const error = await catchErr(assignCertificationOfficerToSession)({ sessionId, certificationOfficerId, jurySessionRepository });
 
       // then
       expect(error).to.be.an.instanceOf(ObjectValidationError);
@@ -38,7 +38,7 @@ describe('Unit | UseCase | assign-certification-officer-to-session', () => {
         certificationOfficerId = 'notANumber';
 
         // when
-        const error = await catchErr(assignCertificationOfficerToSession)({ sessionId, certificationOfficerId, sessionRepository });
+        const error = await catchErr(assignCertificationOfficerToSession)({ sessionId, certificationOfficerId, jurySessionRepository });
 
         // then
         expect(error).to.be.an.instanceOf(ObjectValidationError);
@@ -50,15 +50,15 @@ describe('Unit | UseCase | assign-certification-officer-to-session', () => {
 
       beforeEach(() => {
         certificationOfficerId = 2;
-        sessionRepository.assignCertificationOfficer.withArgs({ id: sessionId, assignedCertificationOfficerId: certificationOfficerId }).resolves(returnedSession);
+        jurySessionRepository.assignCertificationOfficer.withArgs({ id: sessionId, assignedCertificationOfficerId: certificationOfficerId }).resolves(returnedSession);
       });
 
       it('should return the session after assigningUser to it', async () => {
         // when
-        const actualSession = await assignCertificationOfficerToSession({ sessionId, certificationOfficerId, sessionRepository });
+        const actualSession = await assignCertificationOfficerToSession({ sessionId, certificationOfficerId, jurySessionRepository });
 
         // then
-        expect(sessionRepository.assignCertificationOfficer).to.have.been.calledWith({ id: sessionId, assignedCertificationOfficerId: certificationOfficerId });
+        expect(jurySessionRepository.assignCertificationOfficer).to.have.been.calledWith({ id: sessionId, assignedCertificationOfficerId: certificationOfficerId });
         expect(actualSession).to.equal(returnedSession);
       });
     });
